pyproject.toml
[tool.poetry]
name = "termxtract"
version = "0.1.0"
description = ""
authors = ["shubhanker"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
teanga2 = { git = "https://github.com/teangaNLP/teanga2" }

[tool.poetry.group.dev.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.5.39"
mkdocstrings = {extras = ["python"], version = "^0.26.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
